requirements.txt
streamlit
pillow
langchain
langchain-openai
openai
